requirements.txt
annotated-types==0.7.0
anyio==4.9.0
cffi==1.17.1
click==8.2.1
cryptography==45.0.4
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.112.4
h11==0.16.0
idna==3.10
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
python-decouple==3.8
sniffio==1.3.1
starlette==0.38.6
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn==0.34.3
